Update README.md
Sorry if this seems out of place, yesterday I was testing Günther for the first time (which I found awesome BTW) but I spent a lot of time figuring this out. I know this is the bolt docs already, but I found worthwhile to mention this in here for not so experienced users like me. 😅